+++ cmd/buf/internal/command/registry/module/modulelabel/modulelabellist/modulelabellist.go
@@ -27,6 +27,7 @@ import (
 	"github.com/bufbuild/buf/private/buf/bufprint"
 	"github.com/bufbuild/buf/private/bufpkg/bufparse"
 	"github.com/bufbuild/buf/private/bufpkg/bufregistryapi/bufregistryapimodule"
+	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
 )
 
@@ -58,6 +59,9 @@ func NewCommand(
 			},
 		),
 		BindFlags: flags.Bind,
+		ModifyCobra: func(cmd *cobra.Command) error {
+			return bufcli.RegisterFlagCompletionOutputFormat(cmd, formatFlagName)
+		},
 	}
 }